pyproject.toml
[project]
name = "biobench"
version = "0.1.0"
description = "Benchmark for biology-related computer vision tasks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.4.0",
    "jaxtyping>=0.2.33",
    "beartype>=0.18.5",
    "tyro>=0.8.10",
    "torchmetrics>=1.4.1",
    "open-clip-torch>=2.26.1",
    "datasets>=3.4.1",
    "scikit-learn>=1.5.1",
    "requests>=2.32.3",
    "tqdm>=4.66.5",
    "polars>=1.6.0",
    "marimo>=0.11.25",
    "matplotlib>=3.9.2",
    "wilds>=2.0.0",
    "cloudpickle>=3.0.0",
    "timm>=1.0.9",
    "kaggle>=1.6.17",
    "torchvision>=0.19.1",
    "submitit>=1.5.2",
    "pycocotools>=2.0.8",
    "gdown>=5.2.0",
    "safetensors>=0.5.3",
    "einops>=0.8.1",
    "scipy>=1.15.2",
    "mlcroissant>=1.0.17",
    "statsmodels>=0.14.4",
    "whenever>=0.8.3",
    "pillow>=11.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://samuelstevens.me/biobench/"
Issues = "https://github.com/samuelstevens/biobench/issues"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["F722", "E402", "E501"]

[tool.uv]
package = true
default-groups = "all"

[dependency-groups]
dev = [
    "pdoc3>=0.11.1",
    "pytest>=8.3.5",
    "pytest-xdist>=3.6.1",
    "coverage-badge>=1.1.2",
    "pytest-cov>=6.1.1",
    "pytest-json-report>=1.5.0",
    "hypothesis>=6.130.4",
    "pyinstrument>=5.0.1",
    "transformers>=4.51.3",
    "pytest-timeout>=2.3.1",
    "ffmpeg-python>=0.2.0",
    "patdb>=0.1.0",
]
